Rename all instances of DSCP to SQNC, matching case.
