+++ app/api-v1/controllers/Order/index.js
@@ -1,4 +1,4 @@
-const { runProcess } = require('../../../utils/dscp-api')
+const { runProcess } = require('../../../utils/sqnc-api')
 const db = require('../../../db')
 const { validate, mapOrderData } = require('./helpers')
 const identity = require('../../services/identityService')
