+++ app/env.js
@@ -30,8 +30,8 @@ const vars = envalid.cleanEnv(
     PORT: envalid.port({ default: 80, devDefault: 3000 }),
     API_VERSION: envalid.str({ default: version }),
     API_MAJOR_VERSION: envalid.str({ default: 'v1' }),
-    DSCP_API_HOST: envalid.host({ devDefault: 'localhost' }),
-    DSCP_API_PORT: envalid.port({ devDefault: 3001 }),
+    SQNC_API_HOST: envalid.host({ devDefault: 'localhost' }),
+    SQNC_API_PORT: envalid.port({ devDefault: 3001 }),
     LOG_LEVEL: envalid.str({ default: 'info', devDefault: 'debug' }),
     DB_HOST: envalid.host({ devDefault: 'localhost' }),
     DB_PORT: envalid.port({ default: 5432 }),
