+++ docker-compose.yaml
@@ -23,10 +23,10 @@ services:
     environment:
       - POSTGRES_PASSWORD=postgres
       - POSTGRES_USER=postgres
-      - POSTGRES_DB=dscp
+      - POSTGRES_DB=sqnc
 
-  dscp-identity-service:
-    image: digicatapult/dscp-identity-service:latest
+  sqnc-identity-service:
+    image: digicatapult/sqnc-identity-service:latest
     container_name: identity-service
     command: /bin/sh -c "
       sleep 30 &&
@@ -36,17 +36,17 @@ services:
       - 3002:3002
     environment:
       - PORT=3002
-      - API_HOST=dscp-node
+      - API_HOST=sqnc-node
       - DB_HOST=postgres-identity-service
       - DB_PORT=5432
-      - DB_NAME=dscp
+      - DB_NAME=sqnc
       - DB_USERNAME=postgres
       - DB_PASSWORD=postgres
       - SELF_ADDRESS=5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY
       - AUTH_TYPE=${AUTH_TYPE:-NONE}
 
-  dscp-node:
-    image: digicatapult/dscp-node:latest
+  sqnc-node:
+    image: digicatapult/sqnc-node:latest
     container_name: node
     command:
       --base-path /data/
@@ -73,14 +73,14 @@ services:
       - 8080:8080
       - 5001:5001
 
-  dscp-api:
-    image: digicatapult/dscp-api:latest
-    container_name: dscp-api
+  sqnc-api:
+    image: digicatapult/sqnc-api:latest
+    container_name: sqnc-api
     ports:
       - 3001:3001
     environment:
       - PORT=3001
-      - API_HOST=dscp-node
+      - API_HOST=sqnc-node
       - API_PORT=9944
       - USER_URI=//Alice
       - IPFS_HOST=ipfs
