+++ helm/inteli-api/values.yaml
@@ -2,8 +2,8 @@ config:
   port: 80
   # externalOrigin: "http://localhost:3000" # Overrides the server url in the openapi spec
   # externalPathPrefix: "alice/inteli-api" # Path prefix to be applied to served API routes
-  dscpApiHost: dscp-api
-  dscpApiPort: 80
+  sqncApiHost: sqnc-api
+  sqncApiPort: 80
   logLevel: info
   dbName: inteli
   dbPort: 5432
